Refactor: [M3-6276]: NodeBalancerConfigPanel
