+++ packages/manager/src/features/NodeBalancers/NodeBalancerCreate.tsx
@@ -1,4 +1,29 @@
-import { APIError } from '@linode/api-v4/lib/types';
+import * as React from 'react';
+import ActionsPanel from 'src/components/ActionsPanel';
+import Button from 'src/components/Button';
+import Box from 'src/components/core/Box';
+import Accordion from 'src/components/Accordion';
+import Paper from 'src/components/core/Paper';
+import LandingHeader from 'src/components/LandingHeader';
+import TextField from 'src/components/TextField';
+import Typography from 'src/components/core/Typography';
+import SelectRegionPanel from 'src/components/SelectRegionPanel';
+import getAPIErrorFor from 'src/utilities/getAPIErrorFor';
+import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
+import EUAgreementCheckbox from '../Account/Agreements/EUAgreementCheckbox';
+import { CheckoutSummary } from 'src/components/CheckoutSummary/CheckoutSummary';
+import { ConfirmationDialog } from 'src/components/ConfirmationDialog/ConfirmationDialog';
+import { DocumentTitleSegment } from 'src/components/DocumentTitle';
+import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
+import { isEURegion } from 'src/utilities/formatRegion';
+import { NodeBalancerConfigPanel } from './NodeBalancerConfigPanel';
+import { Notice } from 'src/components/Notice/Notice';
+import { sendCreateNodeBalancerEvent } from 'src/utilities/ga';
+import { TagsInput, Tag } from 'src/components/TagsInput/TagsInput';
+import { useGrants, useProfile } from 'src/queries/profile';
+import { useHistory } from 'react-router-dom';
+import { useNodebalancerCreateMutation } from 'src/queries/nodebalancers';
+import { useRegionsQuery } from 'src/queries/regions';
 import {
   append,
   clone,
@@ -8,42 +33,17 @@ import {
   over,
   pathOr,
 } from 'ramda';
-import * as React from 'react';
-import { useHistory } from 'react-router-dom';
-import ActionsPanel from 'src/components/ActionsPanel';
-import Button from 'src/components/Button';
-import { ConfirmationDialog } from 'src/components/ConfirmationDialog/ConfirmationDialog';
-import Typography from 'src/components/core/Typography';
-import { DocumentTitleSegment } from 'src/components/DocumentTitle';
-import { Notice } from 'src/components/Notice/Notice';
-import SelectRegionPanel from 'src/components/SelectRegionPanel';
-import { TagsInput, Tag } from 'src/components/TagsInput/TagsInput';
-import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
-import { isEURegion } from 'src/utilities/formatRegion';
-import { sendCreateNodeBalancerEvent } from 'src/utilities/ga';
-import getAPIErrorFor from 'src/utilities/getAPIErrorFor';
-import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
-import EUAgreementCheckbox from '../Account/Agreements/EUAgreementCheckbox';
-import NodeBalancerConfigPanel from './NodeBalancerConfigPanel';
 import {
   createNewNodeBalancerConfig,
   createNewNodeBalancerConfigNode,
-  NodeBalancerConfigFieldsWithStatus,
   transformConfigsForRequest,
 } from './utils';
 import {
   useAccountAgreements,
   useMutateAccountAgreements,
 } from 'src/queries/accountAgreements';
-import LandingHeader from 'src/components/LandingHeader';
-import { useNodebalancerCreateMutation } from 'src/queries/nodebalancers';
-import Box from 'src/components/core/Box';
-import { CheckoutSummary } from 'src/components/CheckoutSummary/CheckoutSummary';
-import Accordion from 'src/components/Accordion';
-import Paper from 'src/components/core/Paper';
-import TextField from 'src/components/TextField';
-import { useGrants, useProfile } from 'src/queries/profile';
-import { useRegionsQuery } from 'src/queries/regions';
+import type { APIError } from '@linode/api-v4/lib/types';
+import type { NodeBalancerConfigFieldsWithStatus } from './types';
 
 interface NodeBalancerFieldsState {
   label?: string;
